REFACTORING : FriendCard as a separate composable and add mode
- Add mode to change the button behavior
