+++ mobile/src/androidTest/java/ch/epfl/sdp/mobile/test/ui/social/SocialScreenTest.kt
@@ -5,7 +5,6 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import ch.epfl.sdp.mobile.application.ProfileColor
 import ch.epfl.sdp.mobile.test.state.setContentWithLocalizedStrings
 import ch.epfl.sdp.mobile.ui.social.FollowingState
-import ch.epfl.sdp.mobile.ui.social.FriendCard
 import ch.epfl.sdp.mobile.ui.social.Person
 import ch.epfl.sdp.mobile.ui.social.SocialScreen
 import org.junit.Rule
@@ -39,34 +38,6 @@ class SocialScreenTest {
         rule.setContentWithLocalizedStrings { SocialScreen(state = SnapshotSocialScreenState()) }
 
     rule.onNodeWithText(strings.socialFollowingTitle).assertExists()
-  }
-
-  @Test
-  fun card_displayCorrectName() {
-
-    val p =
-        object : Person {
-          override val backgroundColor: ProfileColor = ProfileColor.Pink
-          override val name: String = "Toto"
-          override val emoji: String = ":3"
-        }
-    rule.setContent { FriendCard(person = p) }
-
-    rule.onNodeWithText("Toto").assertExists()
-  }
-
-  @Test
-  fun card_displayCorrectEmoji() {
-
-    val p =
-        object : Person {
-          override val backgroundColor: ProfileColor = ProfileColor.Pink
-          override val name: String = "Toto"
-          override val emoji: String = ":3"
-        }
-    rule.setContent { FriendCard(person = p) }
-
-    rule.onNodeWithText(":3").assertExists()
   }
 
   @Test
