+++ mobile/src/main/java/ch/epfl/sdp/mobile/ui/social/SocialMode.kt
@@ -1,0 +1,11 @@
+package ch.epfl.sdp.mobile.ui.social
+
+/** This enum is used by the composable [FriendCard] to define its behavior */
+enum class SocialMode {
+
+  /** start a game with the player */
+  Play,
+
+  /** Add the selected player in the current user following list */
+  Follow,
+}
